Addressed review comments, changed build file
