+++ src/test/scala/com/databricks/spark/csv/CsvSuite.scala
@@ -15,7 +15,7 @@
  */
 package com.databricks.spark.csv
 
-import java.io.{IOException, File}
+import java.io.File
 
 import org.apache.spark.sql.test._
 import org.apache.spark.sql.types._
@@ -29,51 +29,6 @@ class CsvSuite extends FunSuite {
   val carsAltFile = "src/test/resources/cars-alternative.csv"
   val emptyFile = "src/test/resources/empty.csv"
   val tempEmptyDir = "target/test/empty/"
-
-  /**
-   * This function deletes a file or a directory with everything that's in it. This function is
-   * copied from Spark with minor modifications made to it. See original source at:
-   * github.com/apache/spark/blob/master/core/src/main/scala/org/apache/spark/util/Utils.scala
-   */
-  private def deleteRecursively(file: File) {
-    def listFilesSafely(file: File): Seq[File] = {
-      if (file.exists()) {
-        val files = file.listFiles()
-        if (files == null) {
-          throw new IOException("Failed to list files for dir: " + file)
-        }
-        files
-      } else {
-        List()
-      }
-    }
-
-    if (file != null) {
-      try {
-        if (file.isDirectory) {
-          var savedIOException: IOException = null
-          for (child <- listFilesSafely(file)) {
-            try {
-              deleteRecursively(child)
-            } catch {
-              // In case of multiple exceptions, only last one will be thrown
-              case ioe: IOException => savedIOException = ioe
-            }
-          }
-          if (savedIOException != null) {
-            throw savedIOException
-          }
-        }
-      } finally {
-        if (!file.delete()) {
-          // Delete can also fail if the file simply did not exist
-          if (file.exists()) {
-            throw new IOException("Failed to delete: " + file.getAbsolutePath)
-          }
-        }
-      }
-    }
-  }
 
   test("DSL test") {
     val results = TestSQLContext
@@ -158,9 +113,9 @@ class CsvSuite extends FunSuite {
     assert(cars.schema.fields(2).name == "C2")
   }
 
-  test("INSERT OVERWRITE test") {
+  test("SQL test insert overwrite") {
     // Create a temp directory for table that will be overwritten
-    deleteRecursively(new File(tempEmptyDir))
+    TestUtils.deleteRecursively(new File(tempEmptyDir))
     new File(tempEmptyDir).mkdirs()
     sql(
       s"""
